requirements.txt
bcrypt==3.1.6
cffi==1.12.3
Click==7.0
Flask==1.0.3
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
MarkupSafe==1.1.1
numpy==1.16.3
pandas==0.24.2
Pillow==6.0.0
psycopg2-binary==2.8.2
pycparser==2.19
python-dateutil==2.8.0
pytz==2019.1
scikit-learn==0.21.1
scipy==1.3.0
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.3
transaction==2.4.0
Werkzeug==0.15.4
WTForms==2.2.1
zope.interface==4.6.0
zope.sqlalchemy==1.1
zope.testing==4.7
